Move debug signkey to SGX-specific directory
OE uses a well-known RSA signing key to sign debug SGX enclaves. This is
only used by SGX, move it to an SGX-specific directory and only build it
into SGX hosts.
