+++ host/sgx/signkey.c
